+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unrelease
+
+### Fixed
+
+- Slightly improved error reporting when invoking gcov.
+
 ## 0.5.0 - 2022-12-04
 
 ### Added